krishimandi/requirements.txt
amqp==2.5.2
anyjson==0.3.3
asgiref==3.2.7
billiard==3.6.3.0
celery==4.4.2
certifi==2019.3.9
chardet==3.0.4
click==7.1.1
dash==1.10.0
dash-core-components==1.9.0
dash-html-components==1.0.3
dash-renderer==1.3.0
dash-table==4.6.2
Django==3.0.4
django-celery==3.3.1
django-celery-beat==2.0.0
django-celery-results==1.2.0
django-crispy-forms==1.9.0
django-filter==2.2.0
django-plotly-dash==1.1.5
django-redis==4.11.0
django-timezone-field==4.0
django-verified-email-field==1.6.1
django-workers==0.2.0
dpd-components==0.1.0
Flask==1.1.2
Flask-Compress==1.4.0
future==0.18.2
httplib2==0.12.1
idna==2.8
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.1
kombu==4.6.8
MarkupSafe==1.1.1
numpy==1.18.2
oauth2==1.9.0.post1
pandas==1.0.3
Pillow==7.0.0
plotly==4.6.0
python-crontab==2.4.1
python-dateutil==2.8.1
pytz==2019.3
redis==3.4.1
requests==2.21.0
retrying==1.3.3
six==1.14.0
sqlparse==0.3.1
urllib3==1.24.2
vine==1.3.0
Werkzeug==1.0.1
yahoo-weather==1.0.7
zipp==3.1.0
